+++ spec/Pim/Bundle/BaseConnectorBundle/Reader/ORM/ProductReaderSpec.php
@@ -156,8 +156,8 @@ class ProductReaderSpec extends ObjectBehavior
                         ),
                         'required' => true,
                         'select2'  => true,
-                        'label'    => 'pim_import_export.export.channel.label',
-                        'help'     => 'pim_import_export.export.channel.help'
+                        'label'    => 'pim_base_connector.export.channel.label',
+                        'help'     => 'pim_base_connector.export.channel.help'
                     )
                 )
             )