+++ src/Pim/Bundle/BaseConnectorBundle/Processor/CategoryProcessor.php
@@ -80,8 +80,8 @@ class CategoryProcessor extends TransformerProcessor
             'circularRefsChecked' => array(
                 'type'    => 'switch',
                 'options' => array(
-                    'label' => 'pim_import_export.import.circularRefsChecked.label',
-                    'help'  => 'pim_import_export.import.circularRefsChecked.help'
+                    'label' => 'pim_base_connector.import.circularRefsChecked.label',
+                    'help'  => 'pim_base_connector.import.circularRefsChecked.help'
                 )
             ),
         );