+++ src/Pim/Bundle/BaseConnectorBundle/Processor/CsvSerializer/Processor.php
@@ -131,21 +131,21 @@ abstract class Processor extends AbstractConfigurableStepElement implements
         return array(
             'delimiter' => array(
                 'options' => array(
-                    'label' => 'pim_import_export.export.delimiter.label',
-                    'help'  => 'pim_import_export.export.delimiter.help'
+                    'label' => 'pim_base_connector.export.delimiter.label',
+                    'help'  => 'pim_base_connector.export.delimiter.help'
                 )
             ),
             'enclosure' => array(
                 'options' => array(
-                    'label' => 'pim_import_export.export.enclosure.label',
-                    'help'  => 'pim_import_export.export.enclosure.help'
+                    'label' => 'pim_base_connector.export.enclosure.label',
+                    'help'  => 'pim_base_connector.export.enclosure.help'
                 )
             ),
             'withHeader' => array(
                 'type' => 'switch',
                 'options' => array(
-                    'label' => 'pim_import_export.export.withHeader.label',
-                    'help'  => 'pim_import_export.export.withHeader.help'
+                    'label' => 'pim_base_connector.export.withHeader.label',
+                    'help'  => 'pim_base_connector.export.withHeader.help'
                 )
             ),
         );