+++ src/Pim/Bundle/EnrichBundle/Controller/GroupController.php
@@ -20,7 +20,6 @@ use Pim\Bundle\EnrichBundle\AbstractController\AbstractController;
 use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\EnrichBundle\Form\Handler\GroupHandler;
 use Pim\Bundle\CatalogBundle\Manager\GroupManager;
-use Pim\Bundle\CatalogBundle\Manager\LocaleManager;
 
 /**
  * Group controller
@@ -52,11 +51,6 @@ class GroupController extends AbstractController
     protected $groupForm;
 
     /**
-     * @var LocaleManager
-     */
-    protected $localeManager;
-
-    /**
      * Constructor
      *
      * @param Request                  $request
@@ -69,7 +63,6 @@ class GroupController extends AbstractController
      * @param GroupManager             $groupManager
      * @param GroupHandler             $groupHandler
      * @param Form                     $groupForm
-     * @param LocaleManager            $localeManager
      */
     public function __construct(
         Request $request,
@@ -81,8 +74,7 @@ class GroupController extends AbstractController
         TranslatorInterface $translator,
         GroupManager $groupManager,
         GroupHandler $groupHandler,
-        Form $groupForm,
-        LocaleManager $localeManager
+        Form $groupForm
     ) {
         parent::__construct(
             $request,
@@ -94,10 +86,9 @@ class GroupController extends AbstractController
             $translator
         );
 
-        $this->groupManager  = $groupManager;
-        $this->groupHandler  = $groupHandler;
-        $this->groupForm     = $groupForm;
-        $this->localeManager = $localeManager;
+        $this->groupManager = $groupManager;
+        $this->groupHandler = $groupHandler;
+        $this->groupForm    = $groupForm;
     }
 
     /**
@@ -112,8 +103,7 @@ class GroupController extends AbstractController
     public function indexAction(Request $request)
     {
         return array(
-            'groupTypes' => array_keys($this->groupManager->getTypeChoices(false)),
-            'localeCode' => $this->localeManager->getUserLocale()->getCode()
+            'groupTypes' => array_keys($this->groupManager->getTypeChoices(false))
         );
     }
 
@@ -167,7 +157,6 @@ class GroupController extends AbstractController
 
         return array(
             'form'         => $this->groupForm->createView(),
-            'dataLocale'   => $this->localeManager->getUserLocale()->getCode(),
             'currentGroup' => $group->getId()
         );
     }