+++ src/Pim/Bundle/EnrichBundle/Tests/Unit/MassEditAction/EditCommonAttributesTest.php
@@ -24,12 +24,12 @@ class EditCommonAttributesTest extends \PHPUnit_Framework_TestCase
         $this->locale              = $this->getLocaleMock();
 
         $this->productManager      = $this->getProductManagerMock($this->objectManager, $this->attributeRepository);
-        $this->localeManager       = $this->getLocaleManagerMock($this->locale);
+        $this->userContext         = $this->getUserContextMock($this->locale);
         $this->currencyManager     = $this->getCurrencyManagerMock(array('EUR', 'USD'));
 
         $this->action              = new EditCommonAttributes(
             $this->productManager,
-            $this->localeManager,
+            $this->userContext,
             $this->currencyManager
         );
     }
@@ -218,7 +218,7 @@ class EditCommonAttributesTest extends \PHPUnit_Framework_TestCase
      */
     public function testFormOptions()
     {
-        $this->localeManager
+        $this->userContext
             ->expects($this->any())
             ->method('getUserLocales')
             ->will($this->returnValue(array('fr', 'en', 'pl')));
@@ -321,17 +321,17 @@ class EditCommonAttributesTest extends \PHPUnit_Framework_TestCase
     /**
      * @param mixed $locale
      *
-     * @return \Pim\Bundle\CatalogBundle\Manager\LocaleManager
+     * @return \Pim\Bundle\UserBundle\Context\UserContext
      */
-    protected function getLocaleManagerMock($locale)
+    protected function getUserContextMock($locale)
     {
         $manager = $this
-            ->getMockBuilder('Pim\Bundle\CatalogBundle\Manager\LocaleManager')
+            ->getMockBuilder('Pim\Bundle\UserBundle\Context\UserContext')
             ->disableOriginalConstructor()
             ->getMock();
 
         $manager->expects($this->any())
-            ->method('getLocaleByCode')
+            ->method('getCurrentLocale')
             ->will($this->returnValue($locale));
 
         return $manager;