+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Introduce a BaseConnectorBundle and move readers, processors, writers, archivers and related configuration from ImportExportBundle
 - Introduce a TransformBundle and move cache, converters, encoders, normalizers, transformers and related configuration from ImportExportBundle
 - Renaming of services of ImportExport that have been moved (pim_transform_* and pim_base_connector_*)
+- Move functionality related to user preferences from LocaleManager and ChannelManager to a dedicated UserContext
 
 # 1.0.0-beta-4 - "The Abominable Snow Rabbit" (2014-01-08)